+++ tests/integration/test_card_draw.gd
@@ -1,7 +1,9 @@
 extends "res://tests/UTcommon.gd"
 
 func before_each():
-	setup_board()
+	var confirm_return = setup_board()
+	if confirm_return is GDScriptFunctionState: # Still working.
+		confirm_return = yield(confirm_return, "completed")
 
 
 func test_single_card_draw_use_rectangle():
@@ -147,10 +149,15 @@ func test_draw_multiple_cards_fast():
 
 
 func test_container_custom_card_functions():
+	# warning-ignore:return_value_discarded
 	hand.draw_card()
+	# warning-ignore:return_value_discarded
 	hand.draw_card()
+	# warning-ignore:return_value_discarded
 	hand.draw_card()
+	# warning-ignore:return_value_discarded
 	hand.draw_card()
+	# warning-ignore:return_value_discarded
 	hand.draw_card()
 	var card5: Card = hand.draw_card()
 	yield(yield_to(card5._tween, "tween_all_completed", 1), YIELD)